isolate cie animation component
