+++ ts/screens/authentication/cie/CieCardReaderScreen.tsx
@@ -5,15 +5,13 @@
  */
 import cieManager, { Event as CEvent } from "@pagopa/react-native-cie";
 import * as pot from "italia-ts-commons/lib/pot";
-import { Millisecond } from "italia-ts-commons/lib/units";
-import { Content, Text, View } from "native-base";
+import { Content, Text } from "native-base";
 import * as React from "react";
-import { Animated, Easing, Image, StyleSheet, Vibration } from "react-native";
-import ProgressCircle from "react-native-progress-circle";
+import { StyleSheet, Vibration } from "react-native";
 import { NavigationScreenProps } from "react-navigation";
 import { connect } from "react-redux";
-import AnimatedRing from "../../../components/animations/AnimatedRing";
 import CieNfcOverlay from "../../../components/cie/CieNfcOverlay";
+import CieReadingCardAnimation from "../../../components/cie/CieReadingCardAnimation";
 import { withConditionalView } from "../../../components/helpers/withConditionalView";
 import { ScreenContentHeader } from "../../../components/screens/ScreenContentHeader";
 import TopScreenComponent from "../../../components/screens/TopScreenComponent";
@@ -36,38 +34,13 @@ type NavigationParams = {
 type Props = NavigationScreenProps<NavigationParams> &
   ReturnType<typeof mapStateToProps>;
 
-// Image dimension
-const imgDimension = 180;
-const boxDimension = 245;
-const progressThreshold = 60;
-
 const styles = StyleSheet.create({
   padded: {
     paddingHorizontal: customVariables.contentPadding
-  },
-  imgContainer: {
-    justifyContent: "center",
-    alignItems: "center",
-    height: boxDimension
-  },
-  img: {
-    overflow: "hidden",
-    backgroundColor: customVariables.colorWhite,
-    height: imgDimension - 3,
-    width: imgDimension - 3,
-    borderRadius: imgDimension / 2
-  },
-  rings: {
-    height: boxDimension,
-    width: boxDimension,
-    position: "absolute",
-    overflow: "hidden",
-    justifyContent: "center",
-    alignItems: "center"
   }
 });
 
-enum ReadingState {
+export enum ReadingState {
   "reading" = "reading",
   "error" = "error",
   "completed" = "completed",
@@ -75,7 +48,6 @@ enum ReadingState {
 }
 
 type State = {
-  progressBarValue: number;
   // Get the current status of the card reading
   readingState: ReadingState;
   errorMessage: string;
@@ -85,13 +57,9 @@ type State = {
  *  This screen shown while reading the card
  */
 class CieCardReaderScreen extends React.PureComponent<Props, State> {
-  private progressAnimation: Animated.CompositeAnimation;
-  private progressAnimatedValue: Animated.Value;
-
   constructor(props: Props) {
     super(props);
     this.state = {
-      progressBarValue: 0,
       /* 
       These are the states that can occur when reading the cie (from SDK)
       - waiting_card (we are ready for read ->radar effect)
@@ -102,25 +70,6 @@ class CieCardReaderScreen extends React.PureComponent<Props, State> {
       readingState: ReadingState.waiting_card,
       errorMessage: ""
     };
-    this.progressAnimatedValue = new Animated.Value(0);
-    this.progressAnimatedValue.addListener(anim => {
-      this.setState({ progressBarValue: anim.value });
-    });
-    // Two animation: the first fills the progress with the primary
-    // color up to progressThreshold, the second up to 100
-    // from 0 to 60 in 8 secs
-    const firstAnim = Animated.timing(this.progressAnimatedValue, {
-      toValue: progressThreshold,
-      easing: Easing.linear,
-      duration: 8000
-    });
-    // from 60 to 100 in 10 secs
-    const secondAnim = Animated.timing(this.progressAnimatedValue, {
-      toValue: 100,
-      easing: Easing.linear,
-      duration: 10000
-    });
-    this.progressAnimation = Animated.sequence([firstAnim, secondAnim]);
   }
 
   get ciePin(): string {
@@ -129,30 +78,6 @@ class CieCardReaderScreen extends React.PureComponent<Props, State> {
 
   get cieAuthorizationUri(): string {
     return this.props.navigation.getParam("authorizationUri");
-  }
-
-  private startAnimation = () => {
-    this.progressAnimation.stop();
-    this.setState({ progressBarValue: 0 });
-    this.progressAnimatedValue.setValue(0);
-    this.progressAnimation.start();
-  };
-
-  public componentDidUpdate(_: Props, prevState: State) {
-    // If we start reading the card, start the animation
-    if (
-      prevState.readingState !== ReadingState.reading &&
-      this.state.readingState === ReadingState.reading
-    ) {
-      this.startAnimation();
-    }
-    // If we are not reading the card, stop the animation
-    if (
-      prevState.readingState === ReadingState.reading &&
-      this.state.readingState !== ReadingState.reading
-    ) {
-      this.progressAnimation.stop();
-    }
   }
 
   private handleCieEvent = async (event: CEvent) => {
@@ -195,6 +120,7 @@ class CieCardReaderScreen extends React.PureComponent<Props, State> {
           }
         );
         break;
+
       case "ON_NO_INTERNET_CONNECTION":
         this.setState(
           {
@@ -238,8 +164,6 @@ class CieCardReaderScreen extends React.PureComponent<Props, State> {
   };
 
   private stopCieManager = async () => {
-    this.progressAnimation.stop();
-    this.progressAnimatedValue.removeAllListeners();
     cieManager.removeAllListeners();
     await cieManager.stopListeningNFC();
   };
@@ -265,62 +189,6 @@ class CieCardReaderScreen extends React.PureComponent<Props, State> {
       });
   }
 
-  // TODO: isolate the animation as component
-  private getAnimatedImage = () => {
-    // Setting for 'radar' animation
-    const ringSettings = {
-      dimension: imgDimension,
-      // Three different animation start delays (one is 0), one for each ring
-      delayX1: 700 as Millisecond,
-      delayX2: 1400 as Millisecond,
-      duration: 2100 as Millisecond
-    };
-
-    return (
-      <View style={styles.imgContainer}>
-        {this.state.readingState === ReadingState.waiting_card && (
-          <View style={styles.rings}>
-            <AnimatedRing
-              dimension={ringSettings.dimension}
-              startAnimationAfter={0 as Millisecond}
-              duration={ringSettings.duration}
-              boxDimension={boxDimension}
-            />
-            <AnimatedRing
-              dimension={ringSettings.dimension}
-              startAnimationAfter={ringSettings.delayX1}
-              duration={ringSettings.duration}
-              boxDimension={boxDimension}
-            />
-            <AnimatedRing
-              dimension={ringSettings.dimension}
-              startAnimationAfter={ringSettings.delayX2}
-              duration={ringSettings.duration}
-              boxDimension={boxDimension}
-            />
-          </View>
-        )}
-        <ProgressCircle
-          percent={this.state.progressBarValue}
-          radius={imgDimension / 2}
-          borderWidth={3}
-          color={
-            this.state.readingState === ReadingState.error
-              ? customVariables.brandDanger
-              : customVariables.brandPrimary
-          }
-          shadowColor={customVariables.brandLightGray}
-          bgColor={customVariables.brandLightGray}
-        >
-          <Image
-            source={require("../../../../img/landing/place-card-illustration.png")}
-            style={styles.img}
-          />
-        </ProgressCircle>
-      </View>
-    );
-  };
-
   public render(): React.ReactNode {
     return (
       <TopScreenComponent
@@ -344,7 +212,7 @@ class CieCardReaderScreen extends React.PureComponent<Props, State> {
                 ? I18n.t("authentication.cie.card.readerCardHeader")
                 : I18n.t("authentication.cie.card.layCardMessageHeader")}
           </Text>
-          {this.getAnimatedImage()}
+          <CieReadingCardAnimation readingState={this.state.readingState} />
           <Text style={styles.padded} selectable={true}>
             {this.state.readingState === ReadingState.error
               ? this.state.errorMessage
